Rename wisper_stt to whisper_stt
